+++ packages/cli/src/__fixtures__/simple-project/expect/src/locale/en.ts
@@ -2,8 +2,10 @@ import { Catalog, msg } from "@hi18n/core";
 import type { Vocabulary } from ".";
 
 export default new Catalog<Vocabulary>({
+  "example/dynamic-todo": msg(),
   "example/greeting": msg("Hello, world!"),
   "example/greeting-todo-newtranslation": msg(),
+  "example/greeting-todo2": msg(),
   // "example/greeting-unused": msg("Hello, world! (unused translation)"),
   "example/greeting-todo": msg("Hello, world! (TODO)"),
 });