+++ packages/cli/src/__fixtures__/simple-project/expect/src/locale/index.ts
@@ -3,8 +3,10 @@ import catalogEn from "./en";
 import catalogJa from "./ja";
 
 export type Vocabulary = {
+  "example/dynamic-todo": Message;
   "example/greeting": Message;
   "example/greeting-todo-newtranslation": Message;
+  "example/greeting-todo2": Message;
   // "example/greeting-unused": Message;
   "example/greeting-todo": Message;
 };