+++ packages/cli/src/__fixtures__/simple-project/input/src/index.ts
@@ -1,11 +1,16 @@
-import { getTranslator } from "@hi18n/core";
+import { getTranslator, translationId } from "@hi18n/core";
 import { book } from "./locale";
 
 const { t } = getTranslator(book, "ja");
 t("example/greeting");
 
 // @ts-ignore
+translationId(book, "example/dynamic-todo");
+
+// @ts-ignore
 t("example/greeting-todo");
+
+t.todo("example/greeting-todo2");
 
 // @ts-ignore
 t("example/greeting-todo-newtranslation");