+++ packages/cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add support for `translationId`, `t.todo` and `<Translate.Todo>`
+
 ## 0.1.1
 
 - Fix binary name