+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Internal: add support for `translationId`, `t.todo` and `<Translate.Todo>` (for use in `@hi18n/cli`)
+
 ## 0.1.1
 
 - Fix wrong rule names in the recommended config